version

+++ src/main/java/com/backend/tienda/repository/UsuarioRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.backend.tienda.entity.Usuario;
 
-public interface UsuarioRepository extends JpaRepository<Usuario,Long>{
-	
+public interface UsuarioRepository extends JpaRepository<Usuario,Integer>{
 
 }
